// src/resources/buildings/infrastructure/providers/GetExternalBuildingByIdResponse.ts
export interface GetExternalBuildingByIdResponse {
  cashback?: any;
  minRent: number;
  commissionMinRent: number;
  percentOccupancy: number;
  nonRefundableFee: number;
  status?: any;
  id: string;
  name: string;
  info?: any;
  promos?: any;
  phone: string;
  neighborhood?: any;
  website: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  coordinates: {
    crs: {
      type: string;
      properties: {
        name: string;
      };
    };
    type: string;
    coordinates: number[];
  };
  unitCount: number;
  yearBuilt: number;
  yearRenovated?: number;
  floors: number;
  preferredProvider: number;
  preferredFields: any[];
  shortDescription?: any;
  longDescription?: any;
  leaseLength: string;
  propertyManagerId: string;
  creditCheck?: any;
  secondChance?: any;
  uniqueFeatures?: any;
  foobarScore: number;
  active: boolean;
  maxRent: number;
  minBed: number;
  maxBed: number;
  hasLiveFeed: boolean;
  adminFee: number;
  appFee: number;
  nanoId: string;
  email: string;
  area: string;
  metro: string;
  county: string;
  lastUpdated: Date;
  createdAt: Date;
  floorplans: Array<{
    bathroomCount: number;
    rentAvg: number;
    deposit: number;
    minRent: number;
    maxRent: number;
    name: string;
    bedroomCount: number;
    sqftRange?: any;
    sqftAvg: number;
    rentRange: Array<{
      value: number;
      inclusive: boolean;
    }>;
    availability: Date;
    unitsAvailable: number;
    unitsCount: number;
    providerId: number;
    floorplanPhotos: Array<{
      id: string;
      isActive: boolean;
      url: string;
      rank?: any;
      width?: any;
      height?: any;
      format: string;
      providerId: number;
      floorplanId: string;
      status: string;
    }>;
    utilities: any[];
  }>;
  amenities: any[];
  petPolicies: Array<{
    monthlyFee: number;
    deposit: number;
    applicationFee: number;
    petType: string;
    description: string;
    breedRestriction: string;
    interview: string;
    weight?: number;
    count?: number;
    providerId: number;
    allowed?: any;
  }>;
  parkings: Array<{
    id: number;
    type: string;
    fee?: any;
    description?: any;
    assigned?: any;
    spaces?: any;
    propertyId: string;
    providerId: number;
  }>;
  propertyPhotos: Array<{
    id: string;
    isActive: boolean;
    url: string;
    rank: number;
    alt?: any;
    description?: any;
    width?: any;
    height?: any;
    format: string;
    propertyId: string;
    providerId: number;
    status: string;
    photoSource?: any;
  }>;
}